+++ Src/Parsers/OrderBy.cs
@@ -9,14 +9,14 @@ namespace ODataQuery.Parsers
 {
   static class OrderBy
   {
-    public static Parser<char, bool> Direction =>
+    public static readonly Parser<char, bool> Direction =
       RWS
       .Then(String("asc").WithResult(true)
         .Or(String("desc").WithResult(false)))
       .Optional()
-      .Select(x => x.HasValue ? x.Value : true);
+      .Select(x => !x.HasValue || x.Value);
 
-    public static Parser<char, IEnumerable<(Node node, bool asc)>> Parser =>
+    public static readonly Parser<char, IEnumerable<(Node node, bool asc)>> Parser =
       Map((node, dir) => (node, dir),
           Expression,
           Try(Direction)) // Try -> because Direction consumes whitespace, which makes it fail if it's followed by a comma